+++ urlfilter/decode_test.go
@@ -33,13 +33,15 @@ type Filter struct {
 	Multi    []string
 	MultiNEQ []int
 
-	Time time.Time
+	Time         time.Time
+	StartTimeGTE time.Time
 
 	NullBool    sql.NullBool
 	NullInt64   sql.NullInt64
 	NullFloat64 sql.NullFloat64
 	NullString  sql.NullString
 
+	Map    map[string]string
 	Custom CustomField
 
 	Omit []byte `pg:"-"`
@@ -59,12 +61,18 @@ var _ = Describe("Decode", func() {
 			"multi":      {"one", "two"},
 			"multi__neq": {"3", "4"},
 
-			"time": {"1970-01-01 00:00:00+00:00:00"},
+			"time":            {"1970-01-01 00:00:00+00:00:00"},
+			"start_time__gte": {"1970-01-01 00:00:00+00:00:00"},
 
 			"null_bool":    {"t"},
 			"null_int64":   {"1234"},
 			"null_float64": {"1.234"},
 			"null_string":  {"string"},
+
+			"map[foo]":   {`bar`},
+			"map[hello]": {`world`},
+			"map[]":      {"invalid"},
+			"map][":      {"invalid"},
 
 			"custom": {"custom"},
 		})
@@ -81,6 +89,7 @@ var _ = Describe("Decode", func() {
 		Expect(f.MultiNEQ).To(Equal([]int{3, 4}))
 
 		Expect(f.Time).To(BeTemporally("==", time.Unix(0, 0)))
+		Expect(f.StartTimeGTE).To(BeTemporally("==", time.Unix(0, 0)))
 
 		Expect(f.NullBool.Valid).To(BeTrue())
 		Expect(f.NullBool.Bool).To(BeTrue())
@@ -94,6 +103,7 @@ var _ = Describe("Decode", func() {
 		Expect(f.NullString.Valid).To(BeTrue())
 		Expect(f.NullString.String).To(Equal("string"))
 
+		Expect(f.Map).To(Equal(map[string]string{"foo": "bar", "hello": "world"}))
 		Expect(f.Custom.s).To(Equal("custom"))
 	})