+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 - Client listens on Context.Done while waiting for a connection from the pool and returns an error when context context is cancelled.
 - `Query.Column` does not accept relation name any more. Use `Query.Relation` instead which returns an error if relation does not exist.
 - urlvalues package is renamed to urlfilter.
+- `NewModel` and `AddModel` methods of `HooklessModel` interface were renamed to `NextColumnScanner` and `AddColumnScanner` respectively.
+- `types.F` is deprecated in favor of `types.Ident`.
+- `types.Q` is deprecated in favor of `types.Safe`.
+- `TableName` field is deprecated in favor of `tableName`.
+- `pg:",override"` is deprecated in favor of `pg:",inherit"`.
 
 ## v8